project-root/backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
django-crontab==0.7.1
python-dotenv==1.0.0
pytest==7.3.1
pytest-django==4.5.2
factory-boy==3.2.1
drf-spectacular==0.26.1